apps/api/permissions.py
from rest_framework.permissions import BasePermission
from apps.users.models import User

class IsAdmin(BasePermission):
    def has_permission(self, request, view):
        return (
            request.user.is_authenticated and
            request.user.role == User.Role.ADMIN
        )


class IsManagement(BasePermission):
    def has_permission(self, request, view):
        return (
            request.user.is_authenticated and
            request.user.role in [User.Role.ADMIN, User.Role.LIBRARIAN]
        )
from rest_framework.permissions import BasePermission
from apps.users.models import User


class IsAdminOrLibrarian(BasePermission):

    def has_permission(self, request, view):
        return (
            request.user.is_authenticated and
            request.user.role in [User.Role.ADMIN, User.Role.LIBRARIAN]
        )
